Implementation of the pipeline runner.
The module `atlas.pipeline.runner` implements the logic to execute the commands
defined by a pipeline.

A new setting `search_dir` is used to determine the directory where the rg(1)
commands are executed. If it returns `nil`, the directory is unchanged.

stderr is shared between all processes, so errors will be combined in a single
buffer. This is mostly needed to show messages like invalid regex syntax.

A demo project is added to test the results from rg(1).