Interface implementation.
This commit provides the basic implementation of the main interface.

The interface is defined by two windows:

- The prompt window, where the user can type their search query.

  The instance is closed when the cursor leaves this buffer.

- The results window, which contains the items found by running the query.

  If there are errors from the query (like an invalid regex), the error is
  added to the results window as virtual text lines.

Atlas' instances are launched with:

    require("atlas").open()

The `max_results` setting limits how many results are read from ripgrep.
When the limit is reached, the stdin pipe from ripgrep is closed, so it
can be stopped without providing unnecessary data.

If a new search is started when another one is running, the processes for the
old search receive a `SIGTERM`, and their output will be discarded.

The configuration is moved to its own file, since it is starting to grow
considerably.