Add the fixed-strings operator.
If a specifier is preceded by a `=`, the value will not be interpreted as as
regular expression. This is implemented with the `--fixed-strings` argument in
ripgrep.

The initial replacements (`.` → `\.`, `*` → `.*`) are discarded. To search for
a literal pattern like `.rs`, it is better to use `=.rs`.

`-` and `=` can be combined. For example, `-=/..` finds files without two
consecutive dots.