Improve fold handling with mixed subtree.
Before this patch, the fold marker (`{{{`) was added only if a node had
children. However, this is a problem if a subtree contains both nodes with
children and nodes without them, and the later appears after the former.

For example, given a tree like this:

    a {{{1
      b {{{2
        c
        d
      e
    f

When the fold for `b` is closed, Nvim will include the node for `e`.

The immediate solution is to add a fold marker to `e`:

    a {{{1
      b {{{2
        c
        d
      e {{{2
    f {{{1

This allows Nvim to correctly compute the range for the fold. Unfortunately, the
behaviour of the `za` Normal command is confusing when it is executed on the
line of the `e` node: Nvim closes the fold for `e` (which is empty), and this
action does not have any visible effect, so users may think that it didn't do
anything.

As a workaround, instead of `za`, Atlas now does its own toggle (`zo` and `zc`)
based on the `foldclosed` function. The only difference with the regular `za`
command is that it will tries to close the fold again if it appears to be open.

A better solution could be to set `foldexpr` to a function that is aware on how
the tree is built, but this workaround is enough for now.