History implementation.
Accepted filters are appended to the variable `g:ATLAS_HISTORY`. In the prompt,
the keys <C-Up> and <C-Down> are used to navigate through the history entries.

The initial value for the prompt is initialized with the last value from the
history.

Since the history is stored in a global variable with no lowercase letters, its
value is stored in ShaDa files by Neovim.