Combine content filters.
A prompt with multiple content filters (like `/foo /bar`) will emit entries
only if they match all filters. Before this commit, only the first filter was
used to collect lines from files.

Exclude content filters (like `!/foo`) are applied to the matching lines.
Thus, the lines in the results tree are those matching _all_ content filters
and _none_ of the exclude filters.

This functionality does not exist in ripgrep. To support it, Atlas builds a
complex graph of pipelines, which can be split in three main categories:

- Build the file list.
- Apply content filters.
- Apply exclude content filters.

If the mkfifo(1) program is available, the file list is shared in all ripgrep
programs for content filters.

For the exclude filters, a single ripgrep program is executed with all filters,
and every entry found in the content filters is written to it. If that ripgrep
finds a match, the entry is marked as `exclude`, and will not appear in the
results tree.