Fix performance on large directories.
The luajit profiler (via `require("jit.p").start`) detected that
`vim.fs.basename` and `vim.gsplit` were consuming a lot of CPU time
when the results tree was very large. This commit reduces the impact
by caching the path components, and replacing `vim.fs.basename` with
a custom implementation that does not use regular expressions.

For search programs that emit JSON lines, the reader now stop the
pipeline if the number is results is enough (according to the
`max_results` field in the configuration).